web/requirement.txt
# The webserver
Django

# Provides runserver_plus and shell_plus
django-extensions

# Connects to Postgresql
psycopg2-binary

# Reads .env files
django-dotenv

# used to localize times
pytz

# Used by runserver_plus
Werkzeug

# Converts between units like cups/quarts/milliliters.  This is my fork which adds volume.
git+https://github.com/adamtorres/unit-convert@master#egg=unit-convert

# provides a fancier django admin interface.  Using it because it promised to handle GenericForeignKey nicely.
django-grappelli

# Trying to add colors to print-sql output
pygments

# to parse dates from invoice files
python-dateutil

# mainly for the serializer
djangorestframework
# django-filter for if/when more user-driven filtering is desired.

# For models which reference themselves. ex RawState.next_state
djangorestframework-recursive